% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edx.capstone.movielens.data-package.R
\docType{package}
\name{edx.capstone.movielens.data-package}
\alias{edx.capstone.movielens.data}
\alias{edx.capstone.movielens.data-package}
\title{edx.capstone.movielens.data: edX Data Science: Capstone, MovieLens Datasets}
\description{
Provides datasets for MovieLens project of the edX HarvardX PH125.9x Data Science: Capstone course.
}
\author{
\strong{Maintainer}: Azamat Kurbanaev \email{azkurban@gmail.com}

}
\keyword{internal}
